+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## tip
 
 * FEATURE: fetch tenants from the VictoriaLogs backend and allow selecting a tenant in the datasource settings. See [#475](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/475).
+* FEATURE: enable client side caching and make reliable behavior in QueryBuilder filters. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/357).
+* FEATURE: add compatibility with Grafana 10.x and 11.x by using dynamic component loading for Combobox.
 
 ## v0.22.4